# multisig-wallet/swagger.yaml
openapi: 3.0.3
info:
  title: Welcome to Co-vault Protocol
  description: This is swagger UI for Co-vault protocol.
  version: 1.0.1
servers:
  - url: https://server.covault.xyz/api
...
paths:
...
  /request/getAllRequestList:
    get:
      tags:
        - request
      summary: Returns all request List
      description: Returns all request List
      operationId: getAllRequestList
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getAllRequestList200"
            application/xml:
              schema:
                $ref: "#/components/schemas/getAllRequestList200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
        "422":
          description: Validation exception
  /request/getPsbtFromRequest:
    post:
      tags:
        - request
      summary: Returns specific request by id and pubkey
      description: Returns all request List
      operationId: getPsbtFromRequest
      requestBody:
        description: Get request by vault id and pubkey.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/getPsbtFromRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/getPsbtFromRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/getPsbtFromRequest"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest200"
            application/xml:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
        "422":
          description: Validation exception
  /request/updateRequest:
    post:
      tags:
        - request
      summary: Sign request and broadcasting when threshold get reached
      description: Update reqeust
      operationId: updateRequest
      requestBody:
        description: vaultId of existed valut,  PublicKey of co-signers, minSignCount for Threshold, assets for DAO token, ordinalAddress of updater
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/updateRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/updateRequest"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/updateRequest200"
            application/xml:
              schema:
                $ref: "#/components/schemas/updateRequest200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/updateRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/updateRequest400"
        "422":
          description: Validation exception
  /request/cancelUpdateForRequest:
    post:
      tags:
        - request
      summary: Cancel the request to update the psbt
      description: Cancel reqeust
      operationId: cancelUpdateForRequest
      requestBody:
        description: Cancle the request to update the psbt signing to let other can request to sign that psbt.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/cancelUpdateForRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/cancelUpdateForRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/cancelUpdateForRequest"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest200"
            application/xml:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest400"
        "422":
          description: Validation exception
  /rune/tx-check:
    post:
      tags:
        - rune
      summary: Checking transaction if that is confirmed.
      description: Checking transaction of rune ethcing if that is confirmed or not
      operationId: tx_check
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tx_check200"
            application/xml:
              schema:
                $ref: "#/components/schemas/tx_check200"
        "400":
          description: Invalid sellerInfo, runeTicker and sellPrice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tx_check400"
            application/xml:
              schema:
                $ref: "#/components/schemas/tx_check400"
        "404":
          description: Vault not found
        "422":
          description: Validation exception
  /rune/etch-token:
    post:
      tags:
        - rune
      summary: etching new rune tokens
      description: Etching new rune tokens according to runeName and runeAmount, rune symbol, initialPrice and creator Address
      operationId: etch_token
      requestBody:
        description: buyer Info, inputs Array, listId and psbt
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/etch_runeToken"
          application/xml:
            schema:
              $ref: "#/components/schemas/etch_runeToken"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/etch_runeToken"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/etch_runeToken200"
            application/xml:
              schema:
                $ref: "#/components/schemas/etch_runeToken200"
        "400":
          description: Invalid sellerInfo, runeTicker and sellPrice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/etch_runeToken400"
            application/xml:
              schema:
                $ref: "#/components/schemas/etch_runeToken400"
        "404":
          description: Vault not found
        "422":
          description: Validation exception
components:
  schemas:
...
    Assets:
      required:
        - creatorAddress
        - initialPrice
        - runeAmount
        - runeName
        - runeSymbol
      type: object
      properties:
        creatorAddress:
          type: string
          example: tb1p2vdp0uvc68ng6j56mq4gdkd6jfwhj3qhq2zhp44zxnsgp903d9sqjqlq4r
        initialPrice:
          type: string
          example: 3000
        runeAmount:
          type: string
          example: 40000
        runeName:
          type: string
          example: EIGHTEEN.RUNE.TOKEN.TOKEN
        runeSymbol:
          type: string
          example: ✨
...
    getAllRequestList200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Fetch all request successfully.
        payload:
          type: array
          items:
            type: object
            properties:
              assets:
                $ref: "#/components/schemas/Assets"
              _id:
                type: string
                example: 668787fa55b8778f50981d34
              musigId:
                type: string
                example: 6686bf8b6683b6c56b119f43
              type:
                type: string
                example: VaultUpgrade
              transferAmount:
                type: string
                example: ALL
              destinationAddress:
                type: string
                example: tb1q4xs967j366zfzqzhazvpj555m2j43yvcvrhzhwgg6kgnrkh6av9qf73axx
              creator:
                type: string
                example: tb1p644m7l0fa4wv8x84w6wwhgh5alzpmnkvzxp95n776y6ek8348rwsamh86h
              cosigner:
                type: array
                items:
                  type: string
                example:
                  [
                    03abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a,
                    03481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a86,
                  ]
              signedCosigner:
                type: array
                items:
                  type: string
                example: []
              psbt:
                type: array
                items:
                  type: string
                example:
                  [
                    70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000,
                  ]
              threshold:
                type: number
                example: 2
              pending:
                type: string
                example: ""
              createdAt:
                type: string
                example: 2024-07-05T05:39:17.564Z
    getPsbtFromRequest:
      required:
        - id
        - pubKey
      type: object
      properties:
        id:
          type: string
          example: 6686bf8b6683b6c56b119f43
        pubkey:
          type: string
          example: 03481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a86
    getPsbtFromRequest200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Fetch psbt successfully
        payload:
          type: string
          example: 70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000
    getPsbtFromRequest400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Id or pubkey is missing.
        payload:
          type: object
          example: null
    updateRequest:
      required:
        - psbt
        - id
        - pubkey
      type: object
      properties:
        psbt:
          type: array
          items:
            example: 70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0220203abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a48304502210090abcf53964a34f18bb07b6b99cfb9993f7f03dcc6edd166001e67f254321e6102201a28da118400ef8979c55168cace37ec90094d46107c3fde715e34fb8a4d2f4a01010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000
          xml:
            wrapped: true
        id:
          type: string
          xml:
            wrapped: true
          example: 668791f3051db08c6aed8063
        pubkey:
          type: string
          example: 03abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a
    updateRequest200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Request saved sucessfully
        payload:
          type: array
          items:
            example:
              [
                70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000,
                70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0220203abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a48304502210090abcf53964a34f18bb07b6b99cfb9993f7f03dcc6edd166001e67f254321e6102201a28da118400ef8979c55168cace37ec90094d46107c3fde715e34fb8a4d2f4a01010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000,
              ]
    updateRequest400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Id, pubkey or psbt is missing.
        payload:
          type: object
          example: null
    cancelUpdateForRequest:
      required:
        - psbt
        - id
        - pubkey
      type: object
      properties:
        id:
          type: string
          xml:
            wrapped: true
          example: 668791f3051db08c6aed8063
        pubkey:
          type: string
          example: 03481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a86
    cancelUpdateForRequest200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Cancle request sucessfully
        payload:
          type: object
          example: null
    cancelUpdateForRequest400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: You didn't fetch this psbt
        payload:
          type: object
          example: null
...
    tx_check200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: checked transaction successfully.
        payload:
          type: array
          items:
            type: string
            example: 2862855:197
    tx_check400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Get failure while checking transaction.
        payload:
          type: object
          example: null
    etch_runeToken:
      type: object
      properties:
        runeName:
          type: string
          example: I.CAN.DO.IT.EXACTLY
        runeAmount:
          type: string
          example: 50000
        runeSymbol:
          type: string
          example: $
        initialPrice:
          type: string
          example: 20000
        creatorAddress:
          type: string
          example: tb1pcngsk49thk8e5m2ndfqv9sycltrjr4rx0prwhwr22mujl99y6szqw2kv0f
    etch_runeToken200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Generating PSBT successfully
        payload:
          type: object
          properties:
            txId:
              type: string
              example: ""
            sendBTCTxId:
              type: string
              example: 29716e49b28da526059e63b96b329cf6706fcf146416c77bbf3d676d46fd6ef6
            runeName:
              type: string
              example: I.CAN.DO.IT.EXACTLY
            runeSymbol:
              type: string
              example: $
            initialPrice:
              type: string
              example: 20000
            creatorAddress:
              type: string
              example: tb1pcngsk49thk8e5m2ndfqv9sycltrjr4rx0prwhwr22mujl99y6szqw2kv0f
            runeid:
              type: string
              example: ""
            runeAmount:
              type: string
              example: 50000
            remainAmount:
              type: string
              example: 50000
            psbt:
              type: string
              example: 70736274ff0100dd02000000025b44d45cfefd441ff3bfc492f4a646af323c0fdf3cd205387b7a650fc53abe380200000000ffffffff6e4331de162c138f7a4222dee2d9abe9886e3b919b0f8e155cdc92b2450e770e0100000000ffffffff032202000000000000225120e37a7c5457d8873db854d0af0c2da041ad06419722ab3d10ab1c97604618072da00f000000000000225120531a17f198d1e68d4a9ad82a86d9ba925d794417028570d6a234e08095f16960761f0f0000000000225120e37a7c5457d8873db854d0af0c2da041ad06419722ab3d10ab1c97604618072d000000000001012b2202000000000000225120531a17f198d1e68d4a9ad82a86d9ba925d794417028570d6a234e08095f1696001172007b780761d5d37b642d2ca3a81dccefb15514e8c2489a896711a039b2b11aa690001012b34300f0000000000225120e37a7c5457d8873db854d0af0c2da041ad06419722ab3d10ab1c97604618072d011720abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a00000000
            status:
              type: string
              example: Ready
    etch_runeToken400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Get failure while checking transaction.
        payload:
          type: object
          example: null
...